+++ src/components/Simulering/Simulering.module.scss
@@ -44,26 +44,6 @@
   }
 }
 
-.visTabell {
-  width: 100%;
-
-  button {
-    margin: 0 auto;
-
-    @media (min-width: $a-breakpoint-md) {
-      margin: 0;
-    }
-  }
-
-  :global {
-    .navds-read-more__content {
-      border-left: 0;
-      margin-left: 0;
-      padding-left: 0;
-    }
-  }
-}
-
 .tooltip {
   right: 0.375rem;
   background-color: var(--a-white);
