+++ src/components/Simulering/Simulering.module.scss.d.ts
@@ -1,14 +1,12 @@
-import globalClassNames from '../../style'
+import globalClassNames from '../../style.d'
 declare const classNames: typeof globalClassNames & {
   readonly section: 'section'
   readonly info: 'info'
-  readonly visTabell: 'visTabell'
   readonly buttonRow: 'buttonRow'
   readonly buttonRowElement: 'buttonRowElement'
   readonly buttonRowElement__Right: 'buttonRowElement__Right'
   readonly infoIcon: 'infoIcon'
   readonly infoText: 'infoText'
-  readonly 'navds-read-more__content': 'navds-read-more__content'
   readonly tooltip: 'tooltip'
   readonly tooltipTable: 'tooltipTable'
   readonly tooltipLine: 'tooltipLine'
