+++ src/components/Simulering/Simulering.tsx
@@ -2,7 +2,7 @@ import { useEffect, useRef, useState } from 'react'
 
 import { InformationSquareFillIcon } from '@navikt/aksel-icons'
 import { ChevronLeftCircle, ChevronRightCircle } from '@navikt/ds-icons'
-import { Alert, Button, Heading } from '@navikt/ds-react'
+import { Button } from '@navikt/ds-react'
 import Highcharts, {
   SeriesColumnOptions,
   SeriesOptionsType,
@@ -10,74 +10,54 @@ import Highcharts, {
 } from 'highcharts'
 import HighchartsReact from 'highcharts-react-official'
 
-import { ReadMore } from '../common/ReadMore'
-import { Loader } from '@/components/common/Loader'
 import { TabellVisning } from '@/components/TabellVisning'
 import {
-  useAlderspensjonQuery,
   useGetPersonQuery,
   usePensjonsavtalerQuery,
 } from '@/state/api/apiSlice'
 import {
-  AlderspensjonRequestBody,
+  AlderspensjonResponseBody,
   PensjonsavtalerRequestBody,
 } from '@/state/api/apiSlice.types'
-import {
-  generateAlderspensjonRequestBody,
-  generatePensjonsavtalerRequestBody,
-} from '@/state/api/utils'
+import { generatePensjonsavtalerRequestBody } from '@/state/api/utils'
 import { useAppSelector } from '@/state/hooks'
 import {
   selectCurrentSimulation,
-  selectAfp,
-  selectSamboer,
   selectSamtykke,
 } from '@/state/userInput/selectors'
 
+import { PENSJONSGIVENDE_DATA, SERIES_DEFAULT } from './constants'
 import {
-  COLUMN_WIDTH,
-  PENSJONSGIVENDE_DATA,
-  SERIE_NAME_INNTEKT,
-  SERIE_NAME_AFP,
-  SERIE_NAME_TP,
-  SERIE_NAME_ALDERSPENSJON,
-  SERIE_COLOR_INNTEKT,
-  SERIE_COLOR_AFP,
-  SERIE_COLOR_TP,
-  SERIE_COLOR_ALDERSPENSJON,
-  getChartOptions,
+  getChartDefaults,
   generateXAxis,
-  onPointUnclick,
   onVisFaerreAarClick,
   onVisFlereAarClick,
   processPensjonsberegningArray,
   processPensjonsavtalerArray,
 } from './utils'
+import { getChartOptions, onPointUnclick } from './utils-highcharts'
 
 import styles from './Simulering.module.scss'
 
-export function Simulering() {
-  const { startAlder, startMaaned, uttaksgrad } = useAppSelector(
-    selectCurrentSimulation
-  )
+export function Simulering(props: {
+  alderspensjon?: AlderspensjonResponseBody
+  showAfp: boolean
+  showButtonsAndTable: boolean
+}) {
+  const { alderspensjon, showAfp, showButtonsAndTable } = props
   const harSamtykket = useAppSelector(selectSamtykke)
-  const afp = useAppSelector(selectAfp)
-  const harSamboer = useAppSelector(selectSamboer)
-  const [alderspensjonRequestBody, setAlderspensjonRequestBody] = useState<
-    AlderspensjonRequestBody | undefined
-  >(undefined)
   useState<boolean>(false)
   const [showVisFlereAarButton, setShowVisFlereAarButton] =
     useState<boolean>(false)
   const [showVisFaerreAarButton, setShowVisFaerreAarButton] =
     useState<boolean>(false)
+  const { startAlder, startMaaned } = useAppSelector(selectCurrentSimulation)
   const [pensjonsavtalerRequestBody, setPensjonsavtalerRequestBody] = useState<
     PensjonsavtalerRequestBody | undefined
   >(undefined)
   const [chartOptions, setChartOptions] = useState<Highcharts.Options>(
     getChartOptions(styles, setShowVisFlereAarButton, setShowVisFaerreAarButton)
   )
-  const [isVisTabellOpen, setVisTabellOpen] = useState<boolean>(false)
   const [isPensjonsavtaleFlagVisible, setIsPensjonsavtaleFlagVisible] =
     useState<boolean>(false)
   const chartRef = useRef<HighchartsReact.RefObject>(null)
@@ -90,17 +70,6 @@ export function Simulering() {
       }
     )
 
-  const {
-    data: alderspensjon,
-    isLoading,
-    isError,
-  } = useAlderspensjonQuery(
-    alderspensjonRequestBody as AlderspensjonRequestBody,
-    {
-      skip: !alderspensjonRequestBody,
-    }
-  )
-
   useEffect(() => {
     function onPointUnclickEventHandler(e: Event) {
       onPointUnclick(e, chartRef.current?.chart)
@@ -109,20 +78,6 @@ export function Simulering() {
     return () =>
       document.removeEventListener('click', onPointUnclickEventHandler)
   }, [])
-
-  // Hent alderspensjon + AFP
-  useEffect(() => {
-    const requestBody = generateAlderspensjonRequestBody({
-      afp,
-      sivilstand: person?.sivilstand,
-      harSamboer,
-      foedselsdato: person?.foedselsdato,
-      startAlder: startAlder,
-      startMaaned: startMaaned,
-      uttaksgrad: uttaksgrad,
-    })
-    setAlderspensjonRequestBody(requestBody)
-  }, [afp, person, startAlder, startMaaned, uttaksgrad])
 
   // Hent pensjonsavtaler
   useEffect(() => {
@@ -143,31 +98,16 @@ export function Simulering() {
         setIsPensjonsavtaleFlagVisible
       )
       setChartOptions({
-        chart: {
-          type: 'column',
-          scrollablePlotArea: {
-            minWidth: aarArray.length * COLUMN_WIDTH * 1.6,
-            scrollPositionX: 0,
-          },
-        },
-        xAxis: {
-          categories: aarArray,
-        },
+        ...getChartDefaults(aarArray),
         series: [
           {
-            type: 'column',
-            pointWidth: COLUMN_WIDTH,
-            name: SERIE_NAME_INNTEKT,
-            color: SERIE_COLOR_INNTEKT,
+            ...SERIES_DEFAULT.SERIE_INNTEKT,
             data: [...PENSJONSGIVENDE_DATA].splice(0, aarArray.length),
-          },
-          ...(afp === 'ja_privat'
+          } as SeriesOptionsType,
+          ...(showAfp
             ? [
                 {
-                  type: 'column',
-                  pointWidth: COLUMN_WIDTH,
-                  name: SERIE_NAME_AFP,
-                  color: SERIE_COLOR_AFP,
+                  ...SERIES_DEFAULT.SERIE_AFP,
                   /* c8 ignore next 1 */
                   data: processPensjonsberegningArray(alderspensjon.afpPrivat),
                 } as SeriesOptionsType,
@@ -176,10 +116,7 @@ export function Simulering() {
           ...(isPensjonsavtalerSuccess
             ? [
                 {
-                  type: 'column',
-                  pointWidth: COLUMN_WIDTH,
-                  name: SERIE_NAME_TP,
-                  color: SERIE_COLOR_TP,
+                  ...SERIES_DEFAULT.SERIE_TP,
                   /* c8 ignore next 1 */
                   data: processPensjonsavtalerArray(
                     startAlder - 1,
@@ -191,12 +128,9 @@ export function Simulering() {
               ]
             : []),
           {
-            type: 'column',
-            pointWidth: COLUMN_WIDTH,
-            name: SERIE_NAME_ALDERSPENSJON,
-            color: SERIE_COLOR_ALDERSPENSJON,
+            ...SERIES_DEFAULT.SERIE_ALDERSPENSJON,
             data: processPensjonsberegningArray(alderspensjon.alderspensjon),
-          },
+          } as SeriesOptionsType,
         ],
       })
     }
@@ -204,57 +138,45 @@ export function Simulering() {
 
   return (
     <section className={styles.section}>
-      {isLoading && (
-        <Loader
-          data-testid="loader"
-          size="3xlarge"
-          title="Et øyeblikk, vi beregner pensjonen din"
-        />
-      )}
-      {isError && (
-        <Alert variant="error">
-          <Heading spacing size="small" level="2">
-            TODO PEK-119 feilhåndtering Vi klarte ikke å simulere pensjonen din
-          </Heading>
-        </Alert>
-      )}
       <HighchartsReact
         ref={chartRef}
         highcharts={Highcharts}
         options={chartOptions}
       />
-      <div className={styles.buttonRow}>
-        <div className={styles.buttonRowElement}>
-          {/* c8 ignore next 10 - Dette dekkes av cypress scenario graffHorizontalScroll.cy */}
-          {showVisFaerreAarButton && (
-            <Button
-              icon={<ChevronLeftCircle aria-hidden />}
-              iconPosition="left"
-              size="xsmall"
-              variant="tertiary"
-              onClick={onVisFaerreAarClick}
-            >
-              Færre år
-            </Button>
-          )}
+      {showButtonsAndTable && (
+        <div className={styles.buttonRow}>
+          <div className={styles.buttonRowElement}>
+            {/* c8 ignore next 10 - Dette dekkes av cypress scenario graffHorizontalScroll.cy */}
+            {showVisFaerreAarButton && (
+              <Button
+                icon={<ChevronLeftCircle aria-hidden />}
+                iconPosition="left"
+                size="xsmall"
+                variant="tertiary"
+                onClick={onVisFaerreAarClick}
+              >
+                Færre år
+              </Button>
+            )}
+          </div>
+          <div
+            className={`${styles.buttonRowElement} ${styles.buttonRowElement__Right}`}
+          >
+            {/* c8 ignore next 10 - Dette dekkes av cypress scenario graffHorizontalScroll.cy */}
+            {showVisFlereAarButton && (
+              <Button
+                icon={<ChevronRightCircle aria-hidden />}
+                iconPosition="right"
+                size="xsmall"
+                variant="tertiary"
+                onClick={onVisFlereAarClick}
+              >
+                Flere år
+              </Button>
+            )}
+          </div>
         </div>
-        <div
-          className={`${styles.buttonRowElement} ${styles.buttonRowElement__Right}`}
-        >
-          {/* c8 ignore next 10 - Dette dekkes av cypress scenario graffHorizontalScroll.cy */}
-          {showVisFlereAarButton && (
-            <Button
-              icon={<ChevronRightCircle aria-hidden />}
-              iconPosition="right"
-              size="xsmall"
-              variant="tertiary"
-              onClick={onVisFlereAarClick}
-            >
-              Flere år
-            </Button>
-          )}
-        </div>
-      </div>
+      )}
       {isPensjonsavtaleFlagVisible && (
         <div className={styles.info}>
           <InformationSquareFillIcon
@@ -268,24 +190,12 @@ export function Simulering() {
           </p>
         </div>
       )}
-      <ReadMore
-        name="Tabell av beregningen"
-        header={
-          isVisTabellOpen
-            ? 'Lukk tabell av beregningen'
-            : 'Vis tabell av beregningen'
-        }
-        className={styles.visTabell}
-        open={isVisTabellOpen}
-        onClick={() => {
-          setVisTabellOpen(!isVisTabellOpen)
-        }}
-      >
+      {showButtonsAndTable && (
         <TabellVisning
           series={chartOptions.series as SeriesColumnOptions[]}
           aarArray={(chartOptions?.xAxis as XAxisOptions).categories}
         />
-      </ReadMore>
+      )}
     </section>
   )
 }
