+++ src/components/TabellVisning/TabellVisning.module.scss.d.ts
@@ -1,5 +1,7 @@
 import globalClassNames from '../../style.d'
 declare const classNames: typeof globalClassNames & {
+  readonly visTabell: 'visTabell'
+  readonly 'navds-read-more__content': 'navds-read-more__content'
   readonly table: 'table'
   readonly 'navds-table__expanded-row-content': 'navds-table__expanded-row-content'
   readonly 'navds-table__expandable-row--open': 'navds-table__expandable-row--open'
