+++ src/components/common/Alert/Alert.module.scss
@@ -6,7 +6,7 @@
   border-radius: var(--a-border-radius-medium);
   border-color: var(--ac-alert-info-border, var(--a-border-info));
   padding: var(--a-spacing-5);
-  margin-top: var(--a-spacing-4);
+  margin: var(--a-spacing-4) 0 var(--a-spacing-6) 0;
   align-items: center;
 
   &Wrapper {
