+++ src/mocks/data/alderspensjon/62.json
@@ -130,5 +130,6 @@
       "alder": 77,
       "beloep": 80000
     }
-  ]
+  ],
+  "uttakskravIkkeOppfylt": true
 }
