+++ src/mocks/data/alderspensjon/63.json
@@ -122,5 +122,6 @@
       "alder": 77,
       "beloep": 80000
     }
-  ]
+  ],
+  "uttakskravIkkeOppfylt": true
 }
