+++ src/mocks/data/alderspensjon/64.json
@@ -114,5 +114,6 @@
       "alder": 77,
       "beloep": 80000
     }
-  ]
+  ],
+  "uttakskravIkkeOppfylt": true
 }
