+++ src/mocks/data/alderspensjon/65.json
@@ -106,5 +106,6 @@
       "alder": 77,
       "beloep": 80000
     }
-  ]
+  ],
+  "uttakskravIkkeOppfylt": true
 }
