+++ src/mocks/data/alderspensjon/66.json
@@ -98,5 +98,6 @@
       "alder": 77,
       "beloep": 80000
     }
-  ]
+  ],
+  "uttakskravIkkeOppfylt": true
 }
