+++ src/mocks/data/alderspensjon/67.json
@@ -90,5 +90,6 @@
       "alder": 77,
       "beloep": 80000
     }
-  ]
+  ],
+  "uttakskravIkkeOppfylt": false
 }
