+++ src/mocks/data/alderspensjon/68.json
@@ -82,5 +82,6 @@
       "alder": 77,
       "beloep": 80000
     }
-  ]
+  ],
+  "uttakskravIkkeOppfylt": false
 }
