+++ src/mocks/data/alderspensjon/69.json
@@ -74,5 +74,6 @@
       "alder": 77,
       "beloep": 80000
     }
-  ]
+  ],
+  "uttakskravIkkeOppfylt": false
 }
