+++ src/mocks/data/alderspensjon/70.json
@@ -66,5 +66,6 @@
       "alder": 77,
       "beloep": 80000
     }
-  ]
+  ],
+  "uttakskravIkkeOppfylt": false
 }
