+++ src/mocks/data/alderspensjon/71.json
@@ -58,5 +58,6 @@
       "alder": 77,
       "beloep": 80000
     }
-  ]
+  ],
+  "uttakskravIkkeOppfylt": false
 }
