+++ src/mocks/data/alderspensjon/72.json
@@ -50,5 +50,6 @@
       "alder": 77,
       "beloep": 80000
     }
-  ]
+  ],
+  "uttakskravIkkeOppfylt": false
 }
