+++ src/mocks/data/alderspensjon/73.json
@@ -42,5 +42,6 @@
       "alder": 77,
       "beloep": 80000
     }
-  ]
+  ],
+  "uttakskravIkkeOppfylt": false
 }
