+++ src/mocks/data/alderspensjon/74.json
@@ -34,5 +34,6 @@
       "alder": 77,
       "beloep": 80000
     }
-  ]
+  ],
+  "uttakskravIkkeOppfylt": false
 }
