+++ src/mocks/data/alderspensjon/75.json
@@ -26,5 +26,6 @@
       "alder": 77,
       "beloep": 80000
     }
-  ]
+  ],
+  "uttakskravIkkeOppfylt": false
 }
