+++ src/pages/Step5/Step5Feil.tsx
@@ -4,8 +4,7 @@ import { useNavigate } from 'react-router-dom'
 import { getNesteSide } from '../Step4/utils'
 import { Card } from '@/components/common/Card'
 import { externalUrls } from '@/router/routes'
-import { useGetPersonQuery } from '@/state/api/apiSlice'
-import { apiSlice } from '@/state/api/apiSlice'
+import { apiSlice, useGetPersonQuery } from '@/state/api/apiSlice'
 import { useAppDispatch } from '@/state/hooks'
 import { userInputActions } from '@/state/userInput/userInputReducer'
 import { checkHarSamboer } from '@/utils/sivilstand'
